requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
email-validator==2.2.0
google-generativeai==0.8.4
scikit-learn==1.6.1
numpy==1.26.4
scipy==1.15.2
PyMuPDF==1.25.5
python-docx==1.1.2
requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
threadpoolctl==3.6.0
lxml==5.3.2
tqdm==4.67.1
Flask-Mail 